pre-caching, seems to work
put the API in place, not sure it's ideal but it has the data anyway.
using addresses looks a little weird but the front end can figure it out
from returned data so? also seems kind of duplicate info from totalSupply?

added a cache but bypassed it just to test things working.