allow urls to specify interval and start/end
hook up to backend too, it's a bit clunky but it'll do in a pinch. these
results get cached at the next level at interval specified

TODO some decisions over minutiae like do we want price / liquidity at
beginning of window or end of window (time is currently beginning of window).
the first window is weird if we do beginning of window (it's 0) but it
otherwise makes more sense imo, but I did last for now, either way it's a view
of the windows at about the same time (just shifted up one).

TODO probably need to add in better guards around intervals / times we allow.