initial API doc
this does not track the current API, but is an attempt to document what we
want the API to look like to then go make it happen. it's easier to do now
that we have some data, unsurprisingly.

aside from notes in TODO, some notes on changes:

* remove CMCPrice and totalSupply from token (helps cachability, looks unused, moved to stats)
* remove token object in get pair, just give address for each (up for debate, wanna see usage?)

TODO to dig around the uniswap interface to see about anything we may want to
add that isn't covered well here