require time_start, time_end; add sort facilities
see docs for sorting, it's pretty simple. the default is -liquidityUSD for the
endpoints where it was. we can change the semantics, the popular opinion seems
to be `[+|-]key` or `key[:asc|:desc]` api wise, but open to ideas. it ended up
being a little messy to allow sorting any field, but it seems to work. trying
to read into firebase docs more to see if we can push it all down, but it
doesn't seem easy unless we only have collections for hourly and daily and
then can query against those directly (ie no client side aggregation)