fix up windows to go reverse chronological
this is more likely to align windows as end is more likely to be a time
definitely within our system whereas start is the beginning of time, walking
forward from start was leading to window mismatches when 24h data was compared
to 60 day @24h data that were both displaying varying 24h intervals that did
not align. now they align with 'end' time given and go backwards. the cost of
this is order n to go back over the times in reverse order to do the sums
here, though we could add an index for descending time and not need to absorb
this cost (master, effectively, has this cost already, so we're even but
still), i think this is kind of a hot fix and a more proper fix is to limit
the domain for intervals and aggregate and store only daily and hourly
intervals, which will fix the daily data window mismatch thing (eg we were
comparing 1am yesterday-1am today vs 4am yesterday-4am today, and these didn't
match most often).

in any case, agree this is gross / sucks but it does fix the issue.