initial API doc (#8)
* initial API doc

this does not track the current API, but is an attempt to document what we
want the API to look like to then go make it happen. it's easier to do now
that we have some data, unsurprisingly.

aside from notes in TODO, some notes on changes:

* remove CMCPrice and totalSupply from token (helps cachability, looks unused, moved to stats)
* remove token object in get pair, just give address for each (up for debate, wanna see usage?)

TODO to dig around the uniswap interface to see about anything we may want to
add that isn't covered well here

* update with latest discussion

* add defaults for times

* first pass v1 api

I haven't tested this at all, and is mostly a pass in a hurry, so kinda TODO
on all of this, but this is the basic idea I think, it makes the front end a
lot simpler imo (I mean, some of the methods, specifically) and just maps to
the backend which can be more straight forward now.

* change cache default

* fix pairs sums, totals api

pair sums w/o address specified were all getting mashed together, fix this
from backend.

TODO need to do same for tokens
TODO add tests!

* fix up tokens sums, api doc required fields and sorts

* add testing facilities

TODO add errors, validation & require time/frame stuff

* got first couple tests in

* fix liquidity blank, change uri variable names to address

* passing totals test

* go back to old now tested ways

* add checker again?

* match pair api to doc

* match api doc for query params

* add back default liquidity sort TODO need to update docs and add query parameter

* require time_start, time_end; add sort facilities

see docs for sorting, it's pretty simple. the default is -liquidityUSD for the
endpoints where it was. we can change the semantics, the popular opinion seems
to be `[+|-]key` or `key[:asc|:desc]` api wise, but open to ideas. it ended up
being a little messy to allow sorting any field, but it seems to work. trying
to read into firebase docs more to see if we can push it all down, but it
doesn't seem easy unless we only have collections for hourly and daily and
then can query against those directly (ie no client side aggregation)

* update checker

* fix up windows to go reverse chronological

this is more likely to align windows as end is more likely to be a time
definitely within our system whereas start is the beginning of time, walking
forward from start was leading to window mismatches when 24h data was compared
to 60 day @24h data that were both displaying varying 24h intervals that did
not align. now they align with 'end' time given and go backwards. the cost of
this is order n to go back over the times in reverse order to do the sums
here, though we could add an index for descending time and not need to absorb
this cost (master, effectively, has this cost already, so we're even but
still), i think this is kind of a hot fix and a more proper fix is to limit
the domain for intervals and aggregate and store only daily and hourly
intervals, which will fix the daily data window mismatch thing (eg we were
comparing 1am yesterday-1am today vs 4am yesterday-4am today, and these didn't
match most often).

in any case, agree this is gross / sucks but it does fix the issue.

* make a note

* fix sort bug

* align json delimiter oops