fixes mismatch of volume/liquidity
when clicking on a token or pair page, there was a mismatch between the volume
or liquidity number shown in the chart vs the number shown at the top of the
page.

this was happening due to the window shift thing again, basically when we got
a data point back from the api of the most recent data point and crawled
backwards eg 24 hours, that would be 25 hours ago - since the chart goes back
multiple days, it would get this 25th hour data point and add it within the
bounds of the previous 24h -- this is correct, as viewed from last data point
sub 24h but would be out of bounds of the actual last 24h here in the real
world / the 24h window requested, which is the number that is being read on
the leaderboard - this is quite confusing to read, sorry.

the other issue was that we were overwriting the liquidity number with the
first in a window as opposed to the last, so the liquidity mismatch was again
happening due to this bug with the extra data point in a window (which was at
the beginning). this should be closed now, 2 bugs here get squashed really.

the logic is still pretty much a natural disaster, needs some FEMA, sorry